chore: rename util file to utils.ts & move it's test file
